+++ release/llm_tests/llm_serving_release_tests.py
@@ -1,10 +1,11 @@
-from anyscale_service_utils import start_service
-from typing import Any, List, Optional
-import click
 import os
-import anyscale
-import yaml
+from typing import Any, List, Optional
+
+import click
 import pytest
+import yaml
+
+from anyscale_utils import start_service, get_current_compute_config_name
 
 CLOUD = "serve_release_tests_cloud"
 SERVE_CONFIG_FILE = "standalone_serve_config.yaml"
@@ -16,28 +17,17 @@ def get_applications() -> List[Any]:
     return loaded_llm_config["applications"]
 
 
-def get_current_compute_config_name() -> str:
-    """Get the name of the current compute config."""
-    cluster_id = os.environ["ANYSCALE_CLUSTER_ID"]
-    sdk = anyscale.AnyscaleSDK()
-    cluster = sdk.get_cluster(cluster_id)
-    return anyscale.compute_config.get(
-        name="", _id=cluster.result.cluster_compute_id
-    ).name
+def setup_envs(query_url: str):
+    os.environ["OPENAI_API_BASE"] = query_url
 
 
-def setup_envs():
-    os.environ["ANYSCALE_JOB_CLUSTER_ENV_NAME"] = "standalone"
-    os.environ["OPENAI_API_BASE"] = "standalone"
-
-
+@click.command()
 @click.option("--image-uri", type=str, default=None)
 def main(
     image_uri: Optional[str],
 ):
     applications = get_applications()
     compute_config = get_current_compute_config_name()
-    setup_envs()
 
     with start_service(
         service_name="llm_serving_release_test",
@@ -46,8 +36,9 @@ def main(
         applications=applications,
         working_dir="workloads",
         cloud=CLOUD,
-    ) as service_name:
-        print(f"{service_name=} started")
+    ) as query_url:
+        print(f"Service started: {query_url=}")
+        setup_envs(query_url=query_url)
         pytest.main(
             [
                 "./probes",
