+++ release/llm_tests/probes/openai_client.py
@@ -5,9 +5,11 @@ import openai
 
 DEFAULT_PROBE_HEADERS = {"User-Agent": "rayllm-prober", "Rayllm-Trace": "enable"}
 
+API_KEY = "fake_api_key"
+BASE_URL = os.environ["OPENAI_API_BASE"]
 openai_client = openai.Client(
-    api_key=os.environ["OPENAI_API_KEY"],
-    base_url=os.environ["OPENAI_API_BASE"],
+    api_key=API_KEY,
+    base_url=BASE_URL,
     default_headers=DEFAULT_PROBE_HEADERS,
 )
 
@@ -25,7 +27,7 @@ def create_async_client(
             **headers,
         }
     return openai.AsyncClient(
-        api_key=api_key or os.environ["OPENAI_API_KEY"],
-        base_url=base_url or os.environ["OPENAI_API_BASE"],
+        api_key=api_key or API_KEY,
+        base_url=base_url or BASE_URL,
         default_headers=client_headers,
     )
